% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background.info}
\alias{background.info}
\title{Specifications of background signatures}
\format{
A list with one element for each cell line.
Each element of the list
is in turn a list with the elements \describe{

\item{\code{sig.nbinom.size}}{The negative binomial \code{size} dispersion parameter for the
background signature profile. See \code{\link[stats]{NegBinomial}}.}

\item{\code{count.nbinom.mu}}{The mean of the numbers of mutations in \code{input.spectra}.}

\item{\code{count.nbinom.size}}{The negative binomial \code{size} dispersion parameter for the
numbers of mutations caused by the background signature (i.e. \code{count.nbinom.mu}).
See \code{\link[stats]{NegBinomial}}.}

\item{\code{input.spectra}}{The \code{bg.spectra} used to estimate the background.}
}
}
\source{
\code{background.info}
was estimated from \code{\link{HepG2.background.spectra}} and
\code{\link{MCF10A.background.spectra}}.
}
\usage{
background.info
}
\description{
Specifications of background signatures
}
\examples{
background.info[["HepG2"]]$count.nbinom.mu
background.info[["HepG2"]]$count.nbinom.size
background.info[["HepG2"]]$sig.nbinom.size
background.info[["HepG2"]]$background.sig[1:3, ]



}
\keyword{datasets}
