% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_subtract_signature.R
\name{bisect_n_subtract}
\alias{bisect_n_subtract}
\title{Bisect to find n_subtract where P(no exceed) equals target_prob.}
\usage{
bisect_n_subtract(sig, spectrum, target_prob, tol = 0.5)
}
\arguments{
\item{sig}{Signature profile (numeric vector summing to 1).}

\item{spectrum}{Observed counts per channel (numeric vector).}

\item{target_prob}{Target probability.}

\item{tol}{Bisection tolerance.}
}
\value{
Estimated n_subtract (numeric scalar).
}
\description{
Bisect to find n_subtract where P(no exceed) equals target_prob.
}
\keyword{internal}
