% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareBackgroundSignaturePackageData.R
\name{Defunct.EstimateSignatureFromSpectraLH}
\alias{Defunct.EstimateSignatureFromSpectraLH}
\title{Build a signature for background extraction from a matrix of spectra.}
\usage{
Defunct.EstimateSignatureFromSpectraLH(spectra)
}
\arguments{
\item{spectra}{An \code{\link[ICAMS]{ICAMS}} \code{catalog} with
\code{catalog.type = "counts"}.}
}
\value{
A list with the elements
\enumerate{
\item \code{signature} An \code{\link[ICAMS]{ICAMS}}
\code{catalog} with
\code{catalog.type == "counts.signature"}.
\item \code{log10.counts} Mean log base 10 of the
total counts in \code{spectra}
\item \code{sd.log10.counts.per.base} Standard deviation of
\code{log10.counts.per.base}.
}
}
\description{
This function not only produces a signature, but also an
estimate of the number of mutations usually generated
by the signature and an indication of variability around
that estimate.
}
\details{
Only works on SBS 96 signatures.
}
\keyword{internal}
