% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.spectra}
\alias{example.spectra}
\title{Example spectra of cell lines exposed to cisplatin.}
\format{
A list of \code{\link[ICAMS]{ICAMS}} \code{counts} spectra catalogs.
}
\usage{
example.spectra
}
\description{
Example spectra of cell lines exposed to cisplatin.
}
\examples{
rowSums(example.spectra[["MCF10A.cisplatin"]])[1:3]
}
\keyword{datasets}
