% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeBakgroundInfo.R
\name{MakeBackgroundInfo}
\alias{MakeBackgroundInfo}
\title{Estimate a background signature.}
\usage{
MakeBackgroundInfo(bg.spectra, title = "Background.sig")
}
\arguments{
\item{title}{The name of the single column in the output.}

\item{bg.specta}{The spectra from which to compute the background information.}
}
\value{
An single column \code{\link[ICAMS]{ICAMS}} catalog.
}
\description{
Estimate a background signature.
}
