% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimumSpectrumReconstruction.R
\name{OptimumSpectrumReconstrucion}
\alias{OptimumSpectrumReconstrucion}
\title{Find optimum combination of signatures to generate a spectrum}
\usage{
OptimumSpectrumReconstrucion()
}
\description{
Find optimum combination of signatures to generate a spectrum
}
