% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot1StackedSpectrum.R
\name{Plot1StackedSpectrum}
\alias{Plot1StackedSpectrum}
\title{Plot a spectrum as a stacked bar chart}
\usage{
Plot1StackedSpectrum(
  bg.cat.1,
  target.cat.1,
  background.title = "Background",
  target.title = "Target",
  set.neg.zero = TRUE
)
}
\arguments{
\item{bg.cat.1}{XXXX}

\item{target.title}{XXXXX}

\item{set.neg.zero}{XXXX}

\item{target.cat}{XXXX}

\item{background.tite}{XXXX}
}
\description{
Plot a spectrum as a stacked bar chart
}
