% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_sigs_by_exposure.R
\name{plot_stacked_sigs_by_exposure}
\alias{plot_stacked_sigs_by_exposure}
\title{Plot a spectrum as a stacked bar chart with component signatures}
\usage{
plot_stacked_sigs_by_exposure(sigs, exposures)
}
\arguments{
\item{sigs}{An SBS96 catalog.}

\item{exposure}{A vector of length \code{(ncol(sigs))}.}
}
\description{
Plot a spectrum as a stacked bar chart with component signatures
}
