% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpectraAsSigsWtihUncertainty.R
\name{PlotSpectraAsSigsWithUncertainty}
\alias{PlotSpectraAsSigsWithUncertainty}
\title{Convert spectra to signatures and then plot mean with "error" bars}
\usage{
PlotSpectraAsSigsWithUncertainty(spectra, title = "Mean.as.signature")
}
\arguments{
\item{spectra}{An \code{\link[ICAMS]{ICAMS}} spectrum catalog.
Convert each spectrum to a signature and then compute the
mean.}

\item{title}{The name of the output signature.}
}
\value{
The mean of the spectra as a signature, the
constiuent spectra as signatures, and the y positions of the
arrowheads.
}
\description{
Convert spectra to signatures and then plot mean with "error" bars
}
