% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeparateSignatureAndSpectra.R
\name{SeparateSignatureAndSpectra}
\alias{SeparateSignatureAndSpectra}
\title{Calculate the inferred target signature and inferred background and target components of input spectra}
\usage{
SeparateSignatureAndSpectra(
  spectra,
  bg.sig.info,
  m.opts = NULL,
  start.b.fraction = 0.1,
  sig.name = "Inferred.sig"
)
}
\arguments{
\item{spectra}{The spectra from which to subtract the background,
as a matrix or \code{\link[ICAMS]{ICAMS}} catalog.}

\item{bg.sig.info}{Information about the background signature. See
\code{\link{background.info}}.}

\item{m.opts}{Options to pass to \code{\link[nloptr]{nloptr}}.}

\item{start.b.fraction}{The estimated fraction of the mutations in
\code{spectra} due to the background signature.}

\item{sig.name}{A name for the inferred signature}
}
\description{
Calculate the inferred target signature and inferred background and target components of input spectra
}
