% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeBakgroundInfo.R
\name{MakeBackgroundInfo}
\alias{MakeBackgroundInfo}
\title{Estimate a background signature and its characteristics.}
\usage{
MakeBackgroundInfo(bg.spectra, title = "Background.sig")
}
\arguments{
\item{title}{The name of the single column of the signature in the
\code{background.sig} element of the output.}

\item{bg.specta}{The spectra from which to compute the background information.}
}
\value{
A list with the elements \describe{
\item{\code{background.sig}}{The background signatures as a
single-column \code{\link[ICAMS]{ICAMS}} catalog.}
\item{\code{sig.nbinom.size}}{The negative binomial \code{size} dispersion parameter for the
background signature profile. See \code{\link[stats]{NegBinomial}}.}
\item{\code{count.nbinom.mu}}{The mean of the numbers of mutations in \code{bg.spectra}.}
\item{\code{count.nbinom.size}}{The negative binomial \code{size} dispersion parameter for the
numbers of mutations caused by the background signature (i.e. \code{count.nbinom.mu}).
See \code{\link[stats]{NegBinomial}}.}
\item{\code{input.spectra}}{The \code{bg.spectra} used to estimate the background.}
}
}
\description{
Estimate a background signature and its characteristics.
}
