% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot1StackedSpectrum.R
\name{Plot1StackedSpectrum}
\alias{Plot1StackedSpectrum}
\title{Plot a spectrum as a stacked bar chart}
\usage{
Plot1StackedSpectrum(
  background.spectrum,
  target.spectrum,
  background.title = "Background",
  target.title = "Target",
  set.neg.zero = TRUE
)
}
\arguments{
\item{background.spectrum}{Partial spectra due to a background signature.}

\item{target.spectrum}{Partial spectra due to a target signature.}

\item{background.title}{A title for the legend for the partial
spectra due to a background signature.}

\item{target.title}{A title for the legend for the partial
spectra due to a target signature.}

\item{set.neg.zero}{Sometimes after subtraction a part of the
spectra due to a target signatures is negative. If this
argument is true, set these to 0.}
}
\description{
Plot a spectrum as a stacked bar chart
}
